Add files via upload
For this semester, we have created a Kaggle challenge named “Is the driver at fault?” exclusively for you. Here is the link: https://www.kaggle.com/t/53d1fd0d4cc8422bbdf9de2ea61f7e81.

This challenge is designed based on the Driver Accident tabular Dataset which is provided in the link. Kaggle is a crowd-sourced platform to attract, nurture, train and challenge data scientists from all around the world to solve data science, machine learning and predictive analytics problems.

A driver is at fault or no fault based on several direct and indirect variables specific to the accident and beyond. The dataset has 42 features and almost all of them are categorical. Your aim is to train an ML algorithm that optimizes the evaluation metric i.e. Accuracy. Train.csv file has the training data with Fault Label and input features. Test.csv file has unlabeled data. The detailed description of the challenge and dataset can be found in the aforementioned link.